Starting moving code structure around and removing bad code practices
