+++ package.json
@@ -7,7 +7,7 @@
     "tauri": "tauri"
   },
   "devDependencies": {
-    "@tauri-apps/cli": "^1.5.6"
+    "@tauri-apps/cli": "^1.5.7"
   },
   "dependencies": {
     "@tauri-apps/api": "^1.5.1"
