+++ src-tauri/Cargo.toml
@@ -12,7 +12,7 @@ edition = "2021"
 tauri-build = { version = "1.5.0", features = [] }
 
 [dependencies]
-tauri = { version = "1.5.3", features = [ "updater", "os-all", "shell-open"] }
+tauri = { version = "1.5.0", features = [ "updater", "os-all", "shell-open"] }
 serde = { version = "1.0.193", features = ["derive"] }
 serde_json = "1.0.108"
