+++ src-tauri/src/functions.rs
@@ -4,6 +4,7 @@ use std::path::Path;
 use std::io;
 use std::fs::File;
 use std::io::BufRead;
+use std::path::PathBuf;
 use walkdir::{DirEntry, WalkDir};
 
 
@@ -20,117 +21,6 @@ pub fn hide_windows_console(switch: bool) {
             windows_sys::Win32::System::Console::FreeConsole();
         } else {
             windows_sys::Win32::System::Console::AllocConsole();
-        }
-    }
-}
-
-/*
-
-Regex pattern matches on date and time.
-then gets converted to string.
-
-time and date required to build valid date time string.
-
-*/
-pub fn extract_datetime(log_path: &str) -> String {
-    let re = Regex::new(r"(\d{8}).(\d{6})").unwrap();
-    let regex_captures = re.captures(log_path);
-    // dbg!(&regex_captures);
-    match regex_captures {
-        Some(captures) => {
-            let date_str = captures[1].to_string();
-            let time_str = captures[2].to_string();
-
-            // Parse date and time strings into chrono objects
-            let date = NaiveDate::parse_from_str(&date_str, "%Y%m%d").unwrap();
-            let time = NaiveTime::parse_from_str(&time_str, "%H%M%S").unwrap();
-
-            // Create a combined datetime object
-            let datetime = NaiveDateTime::new(date, time);
-
-            // Format the datetime object into the desired format
-            let formatted_datetime = datetime.format("%Y/%m/%d %H:%M:%S").to_string();
-
-            formatted_datetime
-        }
-        None => {
-            // Handle the case where the regex does not match.
-            log::error!("Could not extract datetime from log path: {}", log_path);
-            String::new()
-        }
-    }
-}
-
-/*
-
-Regex pattern matches on the '\test_env\' | \PTF\ | \AET\ in string.
-Used for confirming whether the returned path is actually correctly pulled from source directory.
-
-*/
-
-
-pub fn extract_info_from_log(log_path: &str) -> Option<(String, u32, String)> {
-    // Open the file for reading
-    if let Ok(file) = File::open(log_path) {
-        // Create a regular expression pattern to match the desired text
-        let re = Regex::new(r"Operation configuration: (\w+(?: \w+)*).*?id: (\d+); Release (\w+)").unwrap();
-
-        for line in io::BufReader::new(file).lines() {
-            if let Ok(line) = line {
-                if let Some(captures) = re.captures(&line) {
-                    if let (Some(testtype), Some(id), Some(release)) = (
-                        captures.get(1),
-                        captures.get(2),
-                        captures.get(3),
-                    ) {
-                        return Some((
-                            testtype.as_str().to_string(),
-                            id.as_str().parse().unwrap(),
-                            release.as_str().to_string(),
-                        ));
-                    }
-                }
-            }
-        }
-
-        eprintln!("Text not found in the file.");
-    } else {
-        eprintln!("Failed to open the file.");
-    }
-
-    // Return a default value when no matches are found
-    Some((
-        "Couldn't determine testtype".to_string(),
-        0, // no match so return default 0
-        "Couldn't determine release".to_string(),
-    ))
-}
-
-
-
-/*
-
-Regex pattern matches on the '\test_env\' | \PTF\ | \AET\ in string.
-Used for confirming whether the returned path is actually correctly pulled from source directory.
-
-*/
-pub fn extract_clnt_string(log_path: &str) -> String {
-    let re = Regex::new(r"CLNT\d+").unwrap();
-    let regex_captures = re.captures(log_path);
-
-    match regex_captures {
-        Some(captures) => {
-            let clnt = captures[0].to_string();
-            // Return the test environment string.
-            clnt
-        }
-        None => {
-            // Handle the case where the regex does not match.
-            log::error!(
-                "Could not find CLNT string in test environment string: {}",
-                log_path
-            );
-            "Could not find CLNT string".to_string()
         }
     }
 }
@@ -159,12 +49,16 @@ pub fn search_for_log(search_info: &crate::structs::AppConfig) -> Result<Vec<Str
     let product_number: String = product_number.to_uppercase();
     let serial_number: String = serial_number.to_uppercase();
     let date_yyyyww: String = date_yyyyww.to_uppercase();
-    let drive_letter: String = drive_letter.to_uppercase();
     let test_suite: String = test_suite.to_uppercase();
 
     // Create the folder path to search.
-    let folder_path = format!("{}\\{}\\{}", drive_letter, folder_location, product_number);
+    // let folder_path = format!("{}\\{}\\{}", drive_letter, folder_location, product_number);
+    let mut folder_path = PathBuf::new();
+    folder_path.push(drive_letter);
+    folder_path.push(folder_location);
+    folder_path.push(product_number);
 
+    dbg!(&folder_path);
     // Create a regular expression to match the log file names.
     let log_pattern = format!(".*{}.*", serial_number);
