+++ src-tauri/tauri.conf.json
@@ -24,7 +24,7 @@
       }
     },
     "bundle": {
-      "active": true,
+      "active": false,
       "targets": "all",
       "identifier": "com.tauri.flixis.findtestlog",
       "copyright": "Tariq Dinmohamed (C) 2023",
